// packages/types/src/interfaces.ts
export type ProductType = "ITEM" | "BLOOK" | "TITLE" | "FONT" | "BANNER" | "OTHER";
// ...
export type PermissionType = "CREATE_REPORTS" | "CHANGE_USERNAME" | "CHANGE_NAME_COLOR_TIER_1" | "EARLY_ACCESS_TIER_1" | "USE_CHAT_COLORS" | "UPLOAD_FILES_SMALL" | "MORE_AUCTIONS_TIER_1" | "MORE_CHAT_BADGE_TIER_1" | "CREATE_GUILDS" | "LESS_AUCTION_TAX" | "CUSTOM_TRADING_TABLE_COLOR" | "CUSTOM_AVATAR" | "CUSTOM_BANNER" | "EARLY_ACCESS_TIER_2" | "UPLOAD_FILES_LARGE" | "CHANGE_NAME_COLOR_TIER_2" | "MORE_AUCTIONS_TIER_2" | "MORE_CHAT_BADGE_TIER_2" | "MUTE_USERS" | "BAN_USERS" | "MANAGE_MESSAGES" | "VIEW_AUDIT" | "BLACKLIST_USERS" | "MANAGE_REPORTS" | "REVERT_AUDIT" | "MANAGE_USER_BLOOKS" | "MANAGE_USER_ITEMS" | "MANAGE_USER_TITLES" | "MANAGE_USER_BANNERS" | "MANAGE_USER_FONTS" | "MANAGE_NEWS_POSTS" | "MANAGE_USER_GROUPS" | "MANAGE_CHAT_ROOMS" | "MANAGE_DATA" | "DELETE_USERS" | "MANAGE_GROUPS";
// ...
export type TransactionStatus = "PENDING" | "COMPLETED" | "FAILED" | "REFUNDED" | "DISPUTED";

export enum CurrencyTypeEnum {
  USD = "USD",
  CRYSTAL = "CRYSTAL"
}
export type CurrencyType = "USD" | "CRYSTAL";
// ...
export type UserSubscriptionStatus = "ACTIVE" | "INACTIVE" | "CANCELED";

export interface Auction {
  id: number;
  type: AuctionType;
  price: number;
  buyItNow: boolean;
  expiresAt: Date;
  createdAt: Date;
  updatedAt: Date;
  blookId: number | null;
  blook?: UserBlook | null;
  itemId: number | null;
  item?: UserItem | null;
  bids?: AuctionBid[];
  sellerId: string;
  seller?: User;
  buyerId: string | null;
  buyer?: User | null;
  delistedAt: Date | null;
}

export interface AuctionBid {
  id: number;
  userId: string;
  auctionId: number;
  amount: number;
  createdAt: Date;
  updatedAt: Date;
  auction?: Auction;
  user?: User;
}

export interface AuditLog {
  id: number;
  action: AuditLogAction;
  model: string;
  relatedTo?: AuditLog[];
  relatedBy?: AuditLog[];
  recordId: number;
  delta: JsonValue | null;
  deltaFields: string[];
  isUndone: boolean;
  undoneById: string | null;
  undoneBy?: User | null;
  actorId: string | null;
  actor?: User | null;
  createdAt: Date;
}

export interface Banner {
  id: number;
  name: string;
  imageId: number;
  priority: number;
  createdAt: Date;
  updatedAt: Date;
  resource?: Resource;
  userBanner?: UserBanner[];
  itemShop?: ItemShop[];
  products?: Product[];
  spinnyWheelRewards?: SpinnyWheelReward[];
}
// ...
export interface Boost {
  id: number;
  userId: string;
  multiplier: number;
  type: BoostType;
  solo: boolean;
  createdAt: Date;
  updatedAt: Date;
  expiresAt: Date;
  user?: User;
}

export interface Emoji {
  id: number;
  name: string;
  imageId: number;
  priority: number;
  createdAt: Date;
  updatedAt: Date;
  resource?: Resource;
}

export interface Font {
  id: number;
  name: string;
  resourceId: number;
  default: boolean;
  priority: number;
  createdAt: Date;
  updatedAt: Date;
  resource?: Resource;
  itemShop?: ItemShop[];
  products?: Product[];
  users?: UserFont[];
  usersUsing?: User[];
  spinnyWheelRewards?: SpinnyWheelReward[];
  titles?: Title[];
}
// ...
export interface IpAddress {
  id: number;
  ipAddress: string;
  createdAt: Date;
  updatedAt: Date;
  users?: User[];
  userIpAddresses?: UserIpAddress[];
  userSubscriptions?: UserSubscription[];
  blacklists?: Blacklist[];
  transactions?: Transaction[];
}
// ...
export interface Message {
  id: string;
  roomId: number;
  content: string;
  mentions: string[];
  color: string | null;
  authorId: string;
  replyingToId: string | null;
  discordMessageId: string | null;
  editedAt: Date | null;
  createdAt: Date;
  updatedAt: Date;
  deletedAt: Date | null;
  author?: User;
  replyingTo?: Message | null;
  replies?: Message[];
  room?: Room;
}

export interface NewsPost {
  id: number;
  title: string;
  content: string;
  imageId: number;
  authorId: string;
  createdAt: Date;
  updatedAt: Date;
  image?: Upload;
  author?: User;
  votes?: UserNewsPostVote[];
}

export interface Pack {
  id: number;
  name: string;
  price: number;
  enabled: boolean;
  hidden: boolean;
  imageId: number;
  iconId: number;
  backgroundId: number;
  ambienceId: number | null;
  priority: number;
  createdAt: Date;
  updatedAt: Date;
  blook?: Blook[];
  image?: Resource;
  icon?: Resource;
  background?: Resource;
  ambience?: Resource | null;
}

export interface Product {
  id: number;
  stripeProductId: string | null;
  stripePriceId: string | null;
  name: string;
  description: string | null;
  price: number;
  subscriptionPrice: number | null;
  imageId: number;
  color1: string;
  color2: string;
  type: ProductType;
  itemId: number | null;
  blookId: number | null;
  titleId: number | null;
  fontId: number | null;
  bannerId: number | null;
  groupId: number | null;
  tokens: number | null;
  diamonds: number | null;
  crystals: number | null;
  isSubscription: boolean | null;
  isQuantityCapped: boolean | null;
  priority: number;
  createdAt: Date;
  updatedAt: Date;
  image?: Resource;
  blook?: Blook | null;
  item?: Item | null;
  title?: Title | null;
  font?: Font | null;
  banner?: Banner | null;
  group?: Group | null;
  subscribers?: UserSubscription[];
  stores?: Store[];
  transactions?: Transaction[];
}

export interface Punishment {
  id: number;
  userId: string;
  type: PunishmentType;
  reason: string;
  expiresAt: Date;
  staffId: string;
  createdAt: Date;
  updatedAt: Date;
  staff?: User;
  user?: User;
  blacklists?: Blacklist[];
}

export interface Rarity {
  id: number;
  name: string;
  color: string;
  animationType: RarityAnimationType;
  experience: number;
  affectedByBooster: boolean;
  imageId: number | null;
  priority: number;
  createdAt: Date;
  updatedAt: Date;
  blook?: Blook[];
  item?: Item[];
  image?: Resource | null;
}

export interface Resource {
  id: number;
  path: string;
  reference: string | null;
  createdAt: Date;
  updatedAt: Date;
  banners?: Banner[];
  blookBackgrounds?: Blook[];
  blooks?: Blook[];
  blookVideos?: Blook[];
  emojis?: Emoji[];
  fonts?: Font[];
  groups?: Group[];
  guilds?: Guild[];
  items?: Item[];
  packIcons?: Pack[];
  packImages?: Pack[];
  packBackgrounds?: Pack[];
  packAmbiences?: Pack[];
  products?: Product[];
  userBanners?: User[];
  rarities?: Rarity[];
}

export interface Room {
  id: number;
  name: string;
  public: boolean;
  createdAt: Date;
  updatedAt: Date;
  messages?: Message[];
  users?: User[];
}

export interface Session {
  id: string;
  userId: string;
  createdAt: Date;
  user?: User;
}
// ...
export interface Title {
  id: number;
  name: string;
  color: string;
  fontId: number;
  bold: boolean;
  italic: boolean;
  underline: boolean;
  priority: number;
  createdAt: Date;
  updatedAt: Date;
  font?: Font;
  itemShop?: ItemShop[];
  products?: Product[];
  users?: UserTitle[];
  usersUsing?: User[];
  spinnyWheelRewards?: SpinnyWheelReward[];
}

export interface Transaction {
  id: string;
  userId: string;
  amount: number;
  quantity: number;
  currency: CurrencyType;
  stripePaymentId: string | null;
  productId: number;
  paymentMethodId: number;
  ipAddressId: number;
  status: TransactionStatus;
  createdAt: Date;
  user?: User;
  product?: Product;
  paymentMethod?: UserPaymentMethod;
  ipAddress?: IpAddress;
}

export interface Upload {
  id: number;
  uploadId: string;
  filename: string;
  userId: string;
  createdAt: Date;
  updatedAt: Date;
  user?: User;
  customAvatars?: User[];
  customBanners?: User[];
  newsPosts?: NewsPost[];
}

export interface User {
  id: string;
  username: string;
  password: string;
  email: string | null;
  emailVerified: boolean;
  avatarId: number | null;
  customAvatarId: number | null;
  bannerId: number | null;
  customBannerId: number | null;
  titleId: number;
  fontId: number;
  color: string;
  tokens: number;
  diamonds: number;
  crystals: number;
  experience: number;
  lastClaimed: Date;
  ipAddressId: number;
  createdAt: Date;
  updatedAt: Date;
  deletedAt: Date | null;
  lastSeen: Date;
  readRulesAt: Date | null;
  stripeCustomerId: string | null;
  isSystem: boolean;
  isAi: boolean;
  subscriptions?: UserSubscription[];
  permissions: PermissionType[];
  bids?: AuctionBid[];
  messages?: Message[];
  rooms?: Room[];
  sessions?: Session[];
  friendedBy?: User[];
  friends?: User[];
  blockedBy?: User[];
  blocked?: User[];
  avatar?: UserBlook | null;
  banner?: Resource | null;
  customAvatar?: Upload | null;
  customBanner?: Upload | null;
  font?: Font;
  title?: Title;
  banners?: UserBanner[];
  initiallyObtainedBlooks?: UserBlook[];
  blooks?: UserBlook[];
  discord?: UserDiscord | null;
  groups?: UserGroup[];
  guild?: UserGuild[];
  guildRequests?: UserGuildRequest[];
  ipAddress?: IpAddress;
  ipAddresses?: UserIpAddress[];
  initiallyObtainedItems?: UserItem[];
  items?: UserItem[];
  OAuth?: UserOAuth[];
  authMethods?: UserWebAuthn[];
  paymentMethods?: UserPaymentMethod[];
  givenPunishments?: Punishment[];
  punishments?: Punishment[];
  settings?: UserSetting | null;
  statistics?: UserStatistic | null;
  titles?: UserTitle[];
  fonts?: UserFont[];
  auctions?: Auction[];
  wonAuctions?: Auction[];
  boosts?: Boost[];
  uploads?: Upload[];
  newsPosts?: NewsPost[];
  votedNewsPosts?: UserNewsPostVote[];
  transactions?: Transaction[];
  auditLogs?: AuditLog[];
  undoneLogs?: AuditLog[];
}

export interface UserBanner {
  id: number;
  userId: string;
  bannerId: number;
  createdAt: Date;
  updatedAt: Date;
  banner?: Banner;
  user?: User;
}

export interface UserBlook {
  id: number;
  userId: string;
  blookId: number;
  sold: boolean;
  shiny: boolean;
  initialObtainerId: string;
  obtainedBy: BlookObtainMethod;
  serial: number | null;
  canSell: boolean;
  canTrade: boolean;
  canAuction: boolean;
  createdAt: Date;
  updatedAt: Date;
  blook?: Blook;
  initialObtainer?: User;
  user?: User;
  auctions?: Auction[];
  avatars?: User[];
}

export interface UserDiscord {
  discordId: string;
  username: string;
  avatar: string | null;
  createdAt: Date;
  updatedAt: Date;
  userId: string;
  user?: User;
}

export interface UserGroup {
  id: number;
  groupId: number;
  userId: string;
  createdAt: Date;
  updatedAt: Date;
  group?: Group;
  user?: User;
}

export interface UserGuild {
  id: number;
  guildId: number;
  userId: string;
  userRole: GuildRole;
  createdAt: Date;
  updatedAt: Date;
  guild?: Guild;
  user?: User;
}

export interface UserGuildRequest {
  id: number;
  guildId: number;
  userId: string;
  createdAt: Date;
  updatedAt: Date;
  guild?: Guild;
  user?: User;
}

export interface UserIpAddress {
  id: number;
  userId: string;
  ipAddressId: number;
  uses: number;
  createdAt: Date;
  updatedAt: Date;
  ipAddress?: IpAddress;
  user?: User;
}

export interface UserItem {
  id: number;
  userId: string;
  itemId: number;
  usesLeft: number;
  initialObtainerId: string;
  obtainedBy: ItemObtainMethod;
  createdAt: Date;
  updatedAt: Date;
  initialObtainer?: User;
  item?: Item;
  user?: User;
  auctions?: Auction[];
}

export interface UserOAuth {
  id: number;
  userId: string;
  type: OAuthType;
  tokenType: string;
  accessToken: string;
  refreshToken: string;
  scope: string;
  expiresAt: Date;
  createdAt: Date;
  updatedAt: Date;
  user?: User;
}

export interface UserWebAuthn {
  id: number;
  userId: string;
  credentialId: string;
  publicKey: string;
  counter: number;
  nickname: string | null;
  createdAt: Date;
  updatedAt: Date;
  user?: User;
}

export interface UserPaymentMethod {
  id: number;
  userId: string;
  paymentMethodId: string;
  lastFour: string;
  cardBrand: string;
  primary: boolean;
  createdAt: Date;
  updatedAt: Date;
  deletedAt: Date | null;
  user?: User;
  transactions?: Transaction[];
}

export interface UserSetting {
  id: string;
  openPacksInstantly: boolean;
  onlyRareSounds: boolean;
  friendRequests: SettingFriendRequest;
  tutorialsCompleted: string[];
  chatColor: string | null;
  otpSecret: string | null;
  user?: User;
}

export interface UserStatistic {
  id: string;
  packsOpened: number;
  messagesSent: number;
  user?: User;
}

export interface UserFont {
  id: number;
  userId: string;
  fontId: number;
  createdAt: Date;
  updatedAt: Date;
  font?: Font;
  user?: User;
}

export interface UserTitle {
  id: number;
  userId: string;
  titleId: number;
  createdAt: Date;
  updatedAt: Date;
  title?: Title;
  user?: User;
}

export interface UserSubscription {
  id: number;
  userId: string;
  stripeSubscriptionId: string;
  productId: number;
  status: UserSubscriptionStatus;
  ipAddressId: number;
  createdAt: Date;
  updatedAt: Date;
  expiresAt: Date | null;
  user?: User;
  product?: Product;
  ipAddress?: IpAddress;
}

export interface UserNewsPostVote {
  id: number;
  userId: string;
  newsPostId: number;
  vote: boolean;
  createdAt: Date;
  updatedAt: Date;
  newsPost?: NewsPost;
  user?: User;
}